+++ R/DefineSpatialClasses.R
@@ -20,11 +20,11 @@ setMethod("spatialSoilSummary", signature("SpatialGridLandscape"), function(obje
   l = object@soillist
   if(length(l)==0) return(NULL)
   firstNoNa = which(!unlist(lapply(l,is.na)))[1]
-  s = do.call(summaryFunction, args=list(object=l[[firstNoNa]],...))
+  s = unlist(do.call(summaryFunction, args=list(object=l[[firstNoNa]],...)))
   sm = data.frame(matrix(NA, nrow=length(l), ncol=length(s)))
   colnames(sm) = names(s)
   for(i in 1:length(l)) {
-    if(!is.na(l[[i]])) sm[i,] = do.call(summaryFunction, args=list(object=l[[i]],...))
+    if(!is.na(l[i])) sm[i,] = unlist(do.call(summaryFunction, args=list(object=l[[i]],...)))
   }
   rownames(sm) = rownames(coordinates(object))
   s = sm