+++ R/spwbgrid.R
@@ -1,5 +1,5 @@
 spwbgrid<-function(y, SpParams, meteo, dates = NULL,
-                  summaryFreq = "years", trackSpecies = numeric(),
+                  summaryFreq = "years",
                   control = defaultControl()) {
 
   #check input
@@ -36,13 +36,20 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
   nDays = length(dates)
   nCells = length(y@forestlist)
   nSummary = sum(table(date.factor)>0)
+  t.df = as.numeric(table(df.int))
+
+  #Determine outlet cells (those without downhill neighbors)
+  outlets = which(unlist(lapply(y@waterQ, sum))==0)
+
+  patchsize = y@forestlist[[1]]$patchsize
 
   #Print information area
   cat("\n------------  spwbgrid ------------\n")
-  cat(paste("Grid cells: ", nCells,", area: ", areaSpatialGrid(y)/10000," ha\n", sep=""))
+  cat(paste("Grid cells: ", nCells,", patchsize: ", patchsize," m2, area: ", nCells*patchsize/10000," ha\n", sep=""))
   cat(paste("Meteorological input class: ", class(meteo),"\n", sep=""))
   cat(paste("Number of days to simulate: ",nDays,"\n", sep=""))
-  cat(paste("Number of summaries: ", nSummary,"\n\n", sep=""))
+  cat(paste("Number of landscape summaries: ", nSummary,"\n", sep=""))
+  cat(paste("Number of outlet cells: ", length(outlets),"\n\n"))
 
   #Set control drainage to false
   control$drainage = FALSE
@@ -65,6 +72,9 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
 
 
   #Output matrices
+  DailyRunoff = matrix(0,nrow = nDays, ncol = length(outlets))
+  colnames(DailyRunoff) = outlets
+  rownames(DailyRunoff) = as.character(dates)
   Runon = matrix(0,nrow=nCells, ncol=nSummary)
   colnames(Runon) = levels(date.factor)[1:nSummary]
   Runoff = Runon
@@ -75,6 +85,9 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
   DeepDrainage = Runon
   SoilEvaporation = Runon
   Transpiration = Runon
+  Psi1 = Runon
+  WTD = Runon
+  Volume = Runon
   LandscapeBalance = data.frame(Precipitation = rep(0, nSummary),
                                 Snow = rep(0, nSummary),
                                 Rain = rep(0, nSummary),
@@ -85,7 +98,13 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
                                 DeepDrainage= rep(0, nSummary))
 
 
-  cat(paste("Running simulations:"))
+  summary_function = function(object, model="SX") {
+    list(Psi1 = soil_psi(object)[1],
+         Volume = sum(soil_water(object, model)),
+         WTD = soil_waterTableDepth(object))
+    }
+
+  cat(paste("Running simulations:\n"))
   for(day in 1:nDays) {
     # cat(paste("Day #", day))
     cat(".")
@@ -142,6 +161,9 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
                      gridPrecipitation, gridRadiation, gridWindSpeed,
                      latitude, elevation, slope, aspect,
                      patchsize)
+
+    summary_day = spatialSoilSummary(y, summary_function, control$soilFunctions)
+    summary_df = summary_day@data
     ifactor = df.int[day]
     Runon[,ifactor] = Runon[,ifactor] + df$WaterBalance$Runon
     Runoff[,ifactor] = Runoff[,ifactor] + df$WaterBalance$Runoff
@@ -152,6 +174,11 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
     DeepDrainage[,ifactor] = DeepDrainage[,ifactor] + df$WaterBalance$DeepDrainage
     SoilEvaporation[,ifactor] = SoilEvaporation[,ifactor] + df$WaterBalance$SoilEvaporation
     Transpiration[,ifactor] = Transpiration[,ifactor] + df$WaterBalance$Transpiration
+    Psi1[,ifactor] = Psi1[,ifactor] + summary_df$Psi1/t.df[ifactor]
+    Volume[,ifactor] = Volume[,ifactor] + summary_df$Volume/t.df[ifactor]
+    WTD[,ifactor] = WTD[,ifactor] + summary_df$WTD/t.df[ifactor]
+
+    DailyRunoff[day,] = df$WaterBalance$Runoff[outlets]*patchsize/1e6 ## Runoff in m3/day
 
     #Landscape balance
     LandscapeBalance$Rain[ifactor]= LandscapeBalance$Rain[ifactor] + sum(df$WaterBalance$Rain, na.rm=T)/nCells
@@ -171,8 +198,11 @@ spwbgrid<-function(y, SpParams, meteo, dates = NULL,
   CellBalance<-list(Rain = Rain, Snow = Snow, Interception = Interception, Runon = Runon, Runoff=Runoff,
                     Infiltration=Infiltration, DeepDrainage = DeepDrainage,
                     SoilEvaporation = SoilEvaporation, Transpiration = Transpiration)
+  CellState<-list(Psi1 = Psi1, Volume = Volume, WTD = WTD)
   l <- list(grid = y@grid, LandscapeBalance = LandscapeBalance,
-            CellBalance = CellBalance)
+            CellBalance = CellBalance,
+            CellState = CellState,
+            DailyRunoff = DailyRunoff)
   class(l)<-c("spwbgrid","list")
   return(l)
 }